Chat members list wasn't updating when new agents were added to the chat. State and handler needed to be shared between the components
